requirements.txt
brax==0.13.0
flax==0.10.7
jax[cuda12]==0.6.2
jaxlib==0.6.2
matplotlib==3.10.6
mujoco-mjx==3.3.5
orbax-checkpoint==0.11.24
playground==0.0.5
gymnasium==1.2.0
